Add header to App and give it basic styling, restructure file tree and update imports in index.js
