--- src/App.css
@@ -1,37 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #09d3ac;
-}
-
-#cat-box {
-  margin:25px 100px 0px 100px;
-  height: 250px;
-  width: 250px;
-  background-color: rebeccapurple;
-}
-#name {
-  color:ghostwhite;
-  font-size:24px;
-}
-#alignment{
-  color:dodgerblue;
-  font-size:22px;
-}
